Update README.md
Inspiration
Nowadays, AI and blockchain are emerging in the tech industry. But they are rarely put into use in our daily lives. ABFaucet wants to utilize these solutions to help people have better sentiments hence a happier life.

What it does
It helps users to have a better dating experience using AI and Blockchain technologies.

How we built it
We first use Twilio for KYC in order to have users have an account. In the meantime, we have users to have tokens registered in order for us to proceed with the blockchain DiD verification to achieve transparency and decentralization. In order to achieve the transparency part, we have come up with a matching algorithm using ML as our AI solution where it is statistical-based and distributive-based and zero-knowledge proof as our blockchain solution.

Challenges we ran into
We ran into challenges like not being able to utilize some existing APIs for us to come up with results so we ended up building our own AI with unique features of being able to date with our chatbox if users do not feel prepared for the real-time dating.

Accomplishments that we're proud of
We created our own matching algorithm.

Firstly, we will have users fill out the features they want in the ideal partners; Secondly, we will based on the results recommend a list of profiles; Then, we will based on the "like" and "dislike" results run a logistic regression model to learn users' behavior; Then, we will have better classification which is personalized for the user and it is constantly changing and iterating just for the user; In addition, we also offer an option for users to "date" with a robot if they are not ready for the real dating matching mechanism to happen.

What we learned
Building this app out is so much fun and we enjoyed it!

What's next for ABFaucet
We hope to build the actual application out on iOS and we wish to get funding to keep it going as we have support from PhD students who are researching blockchain topics and making sure the model is accurate.

Built With
ai
javascript
leo
machine-learning
python
react
smart-contract
solidity
Try it out
 GitHub Repo